refactor(jazz-next): index shared runtime scope plans
